+++ packages/react-api/src/api/SQL.d.ts
@@ -1,13 +1,13 @@
 import { Credentials, ExecuteSQL } from '../types';
 
 export function executeSQL({
-    credentials,
-    query,
-    connection,
-    opts 
-  }: {
-    credentials: Credentials,
-    query: string,
-    connection?: string,
-    opts?: unknown
-  }): ExecuteSQL;
+  credentials,
+  query,
+  connection,
+  opts
+}: {
+  credentials: Credentials;
+  query: string;
+  connection?: string;
+  opts?: unknown;
+}): ExecuteSQL;