+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Rename Credentials exported types [#159](https://github.com/CartoDB/carto-react/pull/159)
 - Improve types [#160](https://github.com/CartoDB/carto-react/pull/160)
 - Adapt Legend widget to work without children [#161](https://github.com/CartoDB/carto-react/pull/161)
 - Update deck.gl version to 8.5.0 [#162](https://github.com/CartoDB/carto-react/pull/162)